backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
httpx
python-dotenv==1.0.1
prisma==0.13.0
psycopg2-binary==2.9.9